refactor: DRY PokemonByType
