+++ src/features/pokemon-detail/__tests__/mocks.ts
@@ -2,7 +2,7 @@ import { vi } from "vitest";
 import { PokemonDetail } from "../domain/entities/PokemonDetail";
 import { EvolutionChain } from "../domain/entities/EvolutionChain";
 import { PokemonStat } from "../domain/value-objects/PokemonStat";
-import { PokemonByType } from "../domain/value-objects/PokemonByType";
+import { PokemonByType } from "../../../shared/domain/value-objects";
 import { PokemonDetailRepository } from "../domain/ports/PokemonDetailRepository";
 
 export const mockBulbasaurStats: PokemonStat[] = [
