+++ src/features/pokemon-list/application/mappers/PokemonListMapper.ts
@@ -1,6 +1,6 @@
 import { PokemonListItem } from "../../domain/entities/PokemonListItem";
 import { PokemonByName } from "../../domain/value-objects/PokemonByName";
-import { PokemonByType } from "../../domain/value-objects/PokemonByType";
+import { PokemonByType } from "../../../../shared/domain/value-objects";
 import { IdGenerator } from "../../../../lib/IdGenerator";
 
 export function mapToDomainList(
