+++ src/features/pokemon-list/domain/ports/PokemonRepository.ts
@@ -1,5 +1,5 @@
 import { PokemonByName } from "../value-objects/PokemonByName.ts";
-import { PokemonByType } from "../value-objects/PokemonByType.ts";
+import { PokemonByType } from "../../../../shared/domain/value-objects";
 import { PokemonType } from "../../../../shared/domain/value-objects/PokemonType";
 export interface PokemonRepository {
   findAllByType(type: PokemonType): Promise<PokemonByType[]>;
