--- src/features/pokemon-list/domain/value-objects/PokemonByType.ts
@@ -1,7 +1,0 @@
-export class PokemonByType {
-  public readonly name: string;
-
-  constructor(name: string) {
-    this.name = name;
-  }
-}
