+++ src/features/pokemon-list/infrastructure/react/hooks/__tests__/setupTests.ts
@@ -1,6 +1,6 @@
 import { vi, beforeEach } from "vitest";
 import { PokemonRepository } from "../../../../domain/ports/PokemonRepository";
-import { PokemonByType } from "../../../../domain/value-objects/PokemonByType";
+import { PokemonByType } from "../../../../../../shared/domain/value-objects";
 import { PokemonByName } from "../../../../domain/value-objects/PokemonByName";
 
 // Global test data and mocks
