+++ src/features/pokemon-list/infrastructure/react/hooks/__tests__/usePokemonList.test.ts
@@ -2,7 +2,7 @@ import { renderHook, waitFor } from "@testing-library/react";
 import { vi, it, expect, beforeEach } from "vitest";
 import usePokemonList from "../usePokemonList";
 import * as reduxHooks from "../../../../../../shared/infrastructure/redux/hooks";
-import { PokemonByType } from "../../../../domain/value-objects/PokemonByType";
+import { PokemonByType } from "../../../../../../shared/domain/value-objects";
 import { PokemonByName } from "../../../../domain/value-objects/PokemonByName";
 import { testData } from "./setupTests";
 import {
