+++ src/shared/domain/value-objects/index.ts
@@ -1,1 +1,2 @@
 export { PokemonType } from "./PokemonType";
+export { PokemonByType } from "./PokemonByType";
